llm_simulation/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
databases
pymongo 
python-dotenv
motor